requirements.txt
strands-agents
strands-agents-tools
strands-agents-builder
strands-agents[anthropic]
strands-agents[bedrock]
strands-agents[a2a]
mcp[cli]
mcp-server-git
fastapi
uvicorn[standard]
python-multipart
python-dotenv
pydantic
pydantic-settings
requests
langfuse
ragas
qdrant-client
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-qdrant
docx2txt
pypdf
boto3
langchain-aws
tiktoken
asyncio
websockets
hypothesis>=6.0.0
pytest>=7.0.0
supabase>=2.0.0
postgrest>=0.10.0
